Make the comment test more complete
So I noticed that this test was incomplete because it might preclude
the use of real code leading up to the `#` character, so I just lazily
put `1 + 1` there and realize that there's an issue with how the 
converter is working. I'm not sure at all what the cause is, because 
it's looking up 1️⃣ => "one" => 1 correctly independently.